Add initial support for QuickCheck error generation
This is a somewhat hacky implementation of using `test.check` to generate new tests cases in the style of QuickCheck. Here we go until at least one individual solves all the existing test cases, and then use `test.check` to generate a new test case that causes that individual to fail. This continues until `test.check` is unable to find a test case that causes the individual to fail after 100 trials.