Changed the rules for generating new test cases
This changes from using `gen/small-integer` to `gen/large-integer` in the range [-1000, 1000]. There were problems when I used `gen/small-integer` where it sometimes throw an exception because it couldn't find a few value that wasn't already being used. Switching to a larger range seems to have fixed that.

I also change things so that it won't generate the same error again, i.e., new errors are unique. This tends to generate test inputs that grow out from 0 in both directions by one. For example, a set of test inputs like `(-4 -3 -2 -1 - 1 2 3 4 5)` is not uncommon.