Hacked in periodic addition of test cases and decimation even if not all solved
This adds two new parameters:

* `gens-to-add-random`
* `gens-to-add-test`

The first is how often we should throw away half the population and replace them with random values. So if that many generations pass since the last time we did a "decimation" pass, then we should do one and reset the counter.

The second is how many generations should pass before we add a test case, regardless of whether we're passing all the tests. 

If we pass all the tests cases then we add a new one, replace half the population, and reset both counters.