Add new thresholds for adding test cases
This adds two new thresholds for adding new tests.

If the average error of the best individual (by total error) is less than or equal to `average-error-threshold`, we will add a new case.

If the proportion of zero errors in the best individual (by total error) is at least `proportion-zero-error-threshold` then we will also add a new test case. Note that the best individual by total error is _not_ necessarily the individual with the most zero errors, so we might want to explore a more sophisticated version of this.